Refactoring LLM calls somewhat.
